+++ ui/CHANGES.md
@@ -1,4 +1,5 @@
 #### Version: 1.1.5-SNAPSHOT (LibGDX 1.9.3)
+- **Changed**: It's now impossible to create `FileTypeFilter` `Rule` without providing at least one extension
 
 #### Version: 1.1.4 (LibGDX 1.9.3)
 - **Added**: `BusyBar` - used to indicate that background work is going on - see `TestBusyBar`